Updated to d3fc@14.0.25
The newer version of d3fc adds style rules to the document head at runtime. However, this approach does not work inside a shadow-dom.
I have raised an issue with d3fc, but the current work-around is to use a local .css file with the necessary d3fc rules in it, that can be imported.